+++ src/main/java/com/ecsa/eokz/repository/compraRepository.java
@@ -1,6 +1,6 @@
 package com.ecsa.eokz.repository;
 
-import com.ecsa.eokz.model.*;
+import com.ecsa.eokz.model.compraModel;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface compraRepository extends JpaRepository <compraModel, Long>{}
+public interface compraRepository extends JpaRepository <compraModel, Integer>{}
